+++ components/admin/Sidebar.tsx
@@ -237,7 +237,7 @@ export default function Sidebar() {
           <span>Logout</span>
         </button>
         <div className="text-center text-xs text-gray-500 mt-2">
-          Next CMS v1.3.0
+          Next CMS v1.3.6
         </div>
       </div>
     </aside>